requirements.txt
# Core ML
numpy==2.4.2
pillow==12.1.1

# Training
torch==2.10.0
torchvision==0.25.0

# Dataset
medmnist==3.0.2

# Metrics / utilities
scikit-learn==1.8.0
tqdm==4.67.3
python-dotenv==1.2.1

# Experiment tracking
wandb==0.25.0
